snap: Base on core24 as per newer qemu
qemu-virgil has a very old stable release, so we want to track edge as
we used to do before.

However the latest qemu-virgil is based on core24 so we should do the
same or we'll end up in ABI / API issues (as various libraries won't
match).

So let's do it, but we also require some more changes though, in
particular:
 - Bump libvirt version (using the old one was almost possible after
   some patching, but qemu is just too new at this point to work with a
   such old version of libvirt, so let's get the newest one we can)
 - drop custom libphodav: just use the archive version
 - Expose the libvirt daemons to the $PATH
 - Adapt pci.ids path to be in the (legacy) expected one by libvirt
 - Adapt some commands and dependencies